requirements.txt
boto3==1.26.63
google-api-python-client~=2.77.0
httplib2~=0.21.0
oauth2client~=4.1.3
pandas==1.4.1
prefect==2.7.10
pyarrow==8.0.0
PyDrive~=1.3.1
requests~=2.28.2
SQLAlchemy==1.4.46
wget==3.2

pyspark~=3.3.2